Add Rubocop and resolve basic offenses
Rubocop has proven extremely valuable in other projects I've worked
on. Even in a relatively small codebase like this, it's best to
keep the code consistent.

For now I've just enabled the cops related to spacing (goodbye
trailing whitespace!) so that merging random commits from the
"treasure hunt" won't be too big of a pain. Once that's all merged
we can go ahead and enable some more. I also added it to travis so
new offenses will cause the build to fail.