Use Camcorder to mock IMAP and SMTP
Fixes #103

Webmock and VCR are only for HTTP so they won't be useful here.
Camcorder is a gem that can be used to record and replay arbitrary
method calls on a given object. This includes basic configuration
for mocking IMAP and SMTP and all the recordings for the existing
specs.