Fix connect not invoking a block
Fixes #121

I removed the meta-programming here because it made the code more
difficult to understand and created a near-invisible issue. When
defining a method via `define_method`, block_given? will return
false even if a block is passed in. You could just look for the
block's existence and do `block.call` but that's not ideal.

This also updates the spec_helper a bit to work with the fact that
an actual block gets passed into mock_client.